Phase 1: move legacy fuzzers and modules into notworking/
Only the Python fuzzing path is actively developed, so everything else is moved
out of the live tree (decisions in TECH_DEBT_PLAN.md). The move set was derived
from a static import-closure of fuzzers/fusil-python-threaded + the test suite;
the complement is provably unreferenced by the kept code (verified: fusil.python
still imports, no kept module references a moved one).

- 15 legacy fuzzer scripts -> fuzzers/notworking/ (fusil-python, -clamav,
  -firefox, -gettext, -gimp, -gstreamer, -imagemagick, -libc-printf, -mplayer,
  -ogg123, -php, -poppler, -vlc, -wizzard, -zzuf).
- Legacy fusil modules -> new fusil/notworking/ subpackage: the file/process
  mangling family (mangle*, auto_mangle, dummy_mangle, incr_mangle*,
  process/mangle), the whole network/ subsystem, x11, c_tools, bits, fixpng,
  zzuf, mockup, terminal_echo, time_watch + process/time_watch, process/attach,
  and linux/syslog. Imports among the moved modules (and in the moved fuzzer
  scripts) are rewritten to fusil.notworking.*; references to kept modules are
  unchanged.

Kept in core (still imported by the Python fuzzer / runtime): file_watch,
linux/cpu_load, bytes_generator, unicode_generator, write_code, and xhost (14
lines, still imported by application.py / replay_python.py on the X11 path;
fully removing that path is a Phase 2 dead-feature task).

Excluded here (handled in Phase 2 with the object-mangling removal):
fusil/python object-mangling samples.

Suite green (the one intermittent failure is the pre-existing flaky randomized
h5py generator test, to be seeded in Phase 6). --help and a generate session
on the ASan build both work. Refs #106.
