+++ fusil/notworking/bits.py
