+++ fusil/notworking/c_tools.py
