+++ fusil/notworking/incr_mangle.py
@@ -1,8 +1,8 @@
 from array import array
 from random import choice, randint
 
-from fusil.incr_mangle_op import OPERATIONS
-from fusil.mangle_agent import MangleAgent
+from fusil.notworking.incr_mangle_op import OPERATIONS
+from fusil.notworking.mangle_agent import MangleAgent
 
 MAX_TRY = 25000
