+++ fusil/notworking/mangle.py
@@ -1,8 +1,8 @@
 from array import array
 from random import choice, randint
 
-from fusil.mangle_agent import MangleAgent
-from fusil.mangle_op import MAX_INCR, SPECIAL_VALUES
+from fusil.notworking.mangle_agent import MangleAgent
+from fusil.notworking.mangle_op import MAX_INCR, SPECIAL_VALUES
 from fusil.tools import minmax
