+++ fusil/notworking/mockup.py
